+++ crates/rust-client/src/store/note_record/mod.rs
@@ -50,7 +50,8 @@ pub use input_note_record::InputNoteRecord;
 pub use output_note_record::OutputNoteRecord;
 pub use states::{
     CommittedNoteState, ConsumedAuthenticatedLocalNoteState, ExpectedNoteState, NoteState,
-    STATE_COMMITTED, STATE_CONSUMED_AUTHENTICATED_LOCAL, STATE_CONSUMED_EXTERNAL,
+    ProcessingAuthenticatedNoteState, ProcessingUnauthenticatedNoteState, STATE_COMMITTED,
+    STATE_CONSUMED_AUTHENTICATED_LOCAL, STATE_CONSUMED_EXTERNAL,
     STATE_CONSUMED_UNAUTHENTICATED_LOCAL, STATE_EXPECTED, STATE_PROCESSING_AUTHENTICATED,
     STATE_PROCESSING_UNAUTHENTICATED, STATE_UNVERIFIED,
 };
@@ -298,6 +299,8 @@ pub enum NoteRecordError {
     NoteNotConsumable(String),
     /// Invalid inclusion proof.
     InvalidInclusionProof,
+    /// Invalid state transition.
+    InvalidStateTransition(String),
     /// Error generated during a state transition.
     StateTransitionError(String),
 }
@@ -310,6 +313,7 @@ impl fmt::Display for NoteRecordError {
             NoteError(err) => write!(f, "Note error: {}", err),
             NoteNotConsumable(msg) => write!(f, "Note not consumable: {}", msg),
             InvalidInclusionProof => write!(f, "Invalid inclusion proof"),
+            InvalidStateTransition(msg) => write!(f, "Invalid state transition: {}", msg),
             StateTransitionError(msg) => write!(f, "State transition error: {}", msg),
         }
     }