+++ crates/rust-client/src/store/note_record/states/committed.rs
@@ -2,6 +2,7 @@ use alloc::string::ToString;
 
 use miden_objects::{
     notes::{NoteId, NoteInclusionProof, NoteMetadata},
+    transaction::TransactionId,
     BlockHeader, Digest,
 };
 
@@ -36,7 +37,7 @@ impl NoteStateHandler for CommittedNoteState {
         Ok(None)
     }
 
-    fn nullifier_received(
+    fn consumed_externally(
         &self,
         nullifier_block_height: u32,
     ) -> Result<Option<NoteState>, NoteRecordError> {
@@ -78,12 +79,26 @@ impl NoteStateHandler for CommittedNoteState {
         ))
     }
 
+    fn transaction_committed(
+        &self,
+        _transaction_id: TransactionId,
+        _block_height: u32,
+    ) -> Result<Option<NoteState>, NoteRecordError> {
+        Err(NoteRecordError::InvalidStateTransition(
+            "Only processing notes can be committed in a local transaction".to_string(),
+        ))
+    }
+
     fn metadata(&self) -> Option<&NoteMetadata> {
         Some(&self.metadata)
     }
 
     fn inclusion_proof(&self) -> Option<&NoteInclusionProof> {
         Some(&self.inclusion_proof)
+    }
+
+    fn consumer_transaction_id(&self) -> Option<&TransactionId> {
+        None
     }
 }