+++ crates/rust-client/src/store/note_record/states/consumed_authenticated_local.rs
@@ -2,6 +2,7 @@ use alloc::string::ToString;
 
 use miden_objects::{
     notes::{NoteId, NoteInclusionProof, NoteMetadata},
+    transaction::TransactionId,
     BlockHeader, Digest,
 };
 
@@ -32,7 +33,7 @@ impl NoteStateHandler for ConsumedAuthenticatedLocalNoteState {
         Ok(None)
     }
 
-    fn nullifier_received(
+    fn consumed_externally(
         &self,
         _nullifier_block_height: u32,
     ) -> Result<Option<NoteState>, NoteRecordError> {
@@ -55,12 +56,26 @@ impl NoteStateHandler for ConsumedAuthenticatedLocalNoteState {
         Err(NoteRecordError::NoteNotConsumable("Note already consumed".to_string()))
     }
 
+    fn transaction_committed(
+        &self,
+        _transaction_id: TransactionId,
+        _block_height: u32,
+    ) -> Result<Option<NoteState>, NoteRecordError> {
+        Err(NoteRecordError::InvalidStateTransition(
+            "Only processing notes can be committed in a local transaction".to_string(),
+        ))
+    }
+
     fn metadata(&self) -> Option<&NoteMetadata> {
         Some(&self.metadata)
     }
 
     fn inclusion_proof(&self) -> Option<&NoteInclusionProof> {
         Some(&self.inclusion_proof)
+    }
+
+    fn consumer_transaction_id(&self) -> Option<&TransactionId> {
+        Some(&self.submission_data.consumer_transaction)
     }
 }